Fix update when quotas are reached
Previously, we attempted to create the service and in case of "service
already exists" response updated the existing service.

This was not working in the case the quota was reached, because the
API returns a quota error before checking if the application exists.

With this commit, we no longer attempt to create the application if
already exists and we update it instead, which does not trigger a quota
error.